fix state location
